Remove Support for Tix
Tix was depreciated in Python 3.6 and removed in Python3.13.
In the original code, it would default to importing Tix and if it could not, it would attempt to import tix from tkinter. Instead, we just load tkinter. Further, we no longer create a TixTk class as it is no longer needed.
Tix has been unsupported for 10 years.