Fix architecture detection on Windows
On Windows, `platform.machine()` returns the identifier of the
*host* architecture, which does not necessarily match the
architecture of the running python process - for example, when
running x86 python under x64 Windows, or when running either
x86 or x64 python under arm64 Windows.

This causes `tkinterdnd2` to pick up incorrect OS-specific
directory (`win-x64` in the former, `win-arm64` in the latter
case), which in turn results in library load error due to
incompatible DLL architecture.

Instead, the architecture of the running process should be
obtained from the `PROCESSOR_ARCHITECTURE` environment variable,
which is automatically set by Windows / WOW subsystem.